refactor: seconds_in_periods->units/datetime
